# noblelift-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Noblelift App API
  version: 1.0.0
  description: |
    REST API для корпоративной системы Noblelift.
    Согласовано с ТЗ и текущим фронтендом. Таймстемпы — Unix ms (int64), camelCase в JSON.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Логин по email/паролю
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление токенов по refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Смена пароля текущим пользователем
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /directories:
    get:
      tags: [Directories]
      summary: Список разделов (дерево или плоско)
      parameters:
        - in: query
          name: flat
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Directory' }
    post:
      tags: [Directories]
      summary: Создать раздел
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DirectoryCreateDnD' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Directory' }
  /directories/{id}:
    get:
      tags: [Directories]
      summary: Детали раздела (дети и документы)
      parameters: [ { $ref: '#/components/parameters/DirectoryId' } ]
      responses:
        '200':
          description: Directory details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DirectoryDetails' }
    patch:
      tags: [Directories]
      summary: Обновить раздел
      parameters: [ { $ref: '#/components/parameters/DirectoryId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DirectoryUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Directory' }
    delete:
      tags: [Directories]
      summary: Удалить раздел
      parameters: [ { $ref: '#/components/parameters/DirectoryId' } ]
      responses:
        '204': { description: No Content }
  /documents:
    get:
      tags: [Documents]
      summary: Поиск документов
      parameters:
        - in: query
          name: directoryId
          schema: { type: integer, format: int64 }
        - in: query
          name: q
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedDocuments' }
    post:
      tags: [Documents]
      summary: Создать документ (версия 1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Карточка документа
      parameters: [ { $ref: '#/components/parameters/DocumentId' } ]
      responses:
        '200':
          description: Document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
    patch:
      tags: [Documents]
      summary: Обновить документ (переименовать/переместить)
      parameters: [ { $ref: '#/components/parameters/DocumentId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
    delete:
      tags: [Documents]
      summary: Удалить документ
      parameters: [ { $ref: '#/components/parameters/DocumentId' } ]
      responses:
        '204': { description: No Content }
  /documents/{id}/versions:
    get:
      tags: [Documents]
      summary: Список версий документа
      parameters: [ { $ref: '#/components/parameters/DocumentId' } ]
      responses:
        '200':
          description: Versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DocumentVersion' }
    post:
      tags: [Documents]
      summary: Загрузить новую версию
      parameters: [ { $ref: '#/components/parameters/DocumentId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentVersionCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentVersion' }
...
  /audit:
    get:
      tags: [Audit]
      summary: Список записей аудита
      parameters:
        - in: query
          name: entity
          schema: { type: string }
        - in: query
          name: entityId
          schema: { type: integer, format: int64 }
        - in: query
          name: actorId
          schema: { type: integer, format: int64 }
        - in: query
          name: dateFrom
          schema: { type: integer, format: int64 }
        - in: query
          name: dateTo
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedAuditLog' }
  /audit/{id}:
    get:
      tags: [Audit]
      summary: Деталь записи аудита
      parameters: [ { $ref: '#/components/parameters/AuditId' } ]
      responses:
        '200':
          description: Audit record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditRecord' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
    Offset:
      in: query
      name: offset
      schema: { type: integer, default: 0, minimum: 0 }
...
    DirectoryId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
    DocumentId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...
    AuditId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }

    # ----- Auth -----
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        rememberMe: { type: boolean, default: false }
    RefreshRequest:
      type: object
      required: [refresh]
      properties:
        refresh: { type: string }
    AuthTokens:
      type: object
      properties:
        access: { type: string }
        refresh: { type: string }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
...
    Directory:
      type: object
      properties:
        id: { type: integer, format: int64 }
        parentId: { type: integer, format: int64, nullable: true }
        name: { type: string }
        accessPolicy: { type: object, additionalProperties: true, nullable: true }
    DirectoryCreateDnD:
      type: object
      required: [name]
      properties:
        parentId: { type: integer, format: int64, nullable: true }
        name: { type: string }
        accessPolicy: { type: object, additionalProperties: true, nullable: true }
    DirectoryUpdate:
      type: object
      properties:
        parentId: { type: integer, format: int64, nullable: true }
        name: { type: string }
        accessPolicy: { type: object, additionalProperties: true, nullable: true }
    DirectoryDetails:
      type: object
      properties:
        directory: { $ref: '#/components/schemas/Directory' }
        children:
          type: array
          items: { $ref: '#/components/schemas/Directory' }
        documents:
          type: array
          items: { $ref: '#/components/schemas/Document' }
    Document:
      type: object
      properties:
        id: { type: integer, format: int64 }
        directoryId: { type: integer, format: int64 }
        title: { type: string }
        fileKey: { type: string }
        mime: { type: string, nullable: true }
        size: { type: integer, nullable: true }
        version: { type: integer }
        createdAt: { type: integer, format: int64 }
        createdBy: { type: integer, format: int64 }
    DocumentCreate:
      type: object
      required: [directoryId, title, fileKey]
      properties:
        directoryId: { type: integer, format: int64 }
        title: { type: string }
        fileKey: { type: string }
        mime: { type: string, nullable: true }
        size: { type: integer, nullable: true }
    DocumentUpdate:
      type: object
      properties:
        directoryId: { type: integer, format: int64 }
        title: { type: string }
    DocumentVersion:
      type: object
      properties:
        id: { type: integer, format: int64 }
        documentId: { type: integer, format: int64 }
        version: { type: integer }
        fileKey: { type: string }
        createdAt: { type: integer, format: int64 }
        createdBy: { type: integer, format: int64 }
    DocumentVersionCreate:
      type: object
      required: [fileKey]
      properties:
        fileKey: { type: string }
...
    PaginatedDocuments:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Document' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }

    # ----- Vehicles -----
...
    AuditRecord:
      type: object
      properties:
        id: { type: integer, format: int64 }
        actorId: { type: integer, format: int64, nullable: true }
        action: { type: string }
        entity: { type: string }
        entityId: { type: integer, format: int64 }
        diff: { type: object, additionalProperties: true, nullable: true }
        createdAt: { type: integer, format: int64 }
    PaginatedAuditLog:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AuditRecord' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }